requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu112
dvc==2.58.2
matplotlib
dvc-ssh==2.22.2
efficientnet==1.0.0
efficientnet-pytorch==0.7.1
fire==0.5.0
keras==2.13.1
Keras-Applications==1.0.8
mlflow==2.10.2
numpy==1.24.3
nuscenes-devkit==1.1.11
opencv-python==4.8.0.76
pandas==2.1.1
Pillow==10.1.0
scikit-learn==1.3.2
segmentation-models==1.0.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorflow==2.13.0
torch==2.0.0 --index-url https://download.pytorch.org/whl/cu112
torchvision==0.15.1 --index-url https://download.pytorch.org/whl/cu112
pkbar==0.5
paramiko==3.4.0
pysftp==0.2.9
